Cache attribute information
Tests show the caching for ``GetAreaName`` goes from 4.28 seconds down to 0.03 seconds.
Tests show the caching for ``GetActionName`` goes from 1.52 seconds down to 0.05 seconds.

(These tests are for 1 million calls and are simple basic performance on my machine using r# test runner)